+++ Controller/RegistrationController.php
@@ -18,7 +18,7 @@ use FOS\UserBundle\Form\Factory\FactoryInterface;
 use FOS\UserBundle\FOSUserEvents;
 use FOS\UserBundle\Model\UserInterface;
 use FOS\UserBundle\Model\UserManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -33,7 +33,7 @@ use Symfony\Component\Security\Core\Exception\AccessDeniedException;
  * @author Thibault Duplessis <thibault.duplessis@gmail.com>
  * @author Christophe Coevoet <stof@notk.org>
  */
-class RegistrationController extends Controller
+final class RegistrationController extends AbstractController
 {
     private $eventDispatcher;
     private $formFactory;
@@ -49,8 +49,6 @@ class RegistrationController extends Controller
     }
 
     /**
-     * @param Request $request
-     *
      * @return Response
      */
     public function registerAction(Request $request)
@@ -95,9 +93,9 @@ class RegistrationController extends Controller
             }
         }
 
-        return $this->render('@FOSUser/Registration/register.html.twig', array(
+        return $this->render('@FOSUser/Registration/register.html.twig', [
             'form' => $form->createView(),
-        ));
+        ]);
     }
 
     /**
@@ -118,16 +116,15 @@ class RegistrationController extends Controller
             return new RedirectResponse($this->container->get('router')->generate('fos_user_security_login'));
         }
 
-        return $this->render('@FOSUser/Registration/check_email.html.twig', array(
+        return $this->render('@FOSUser/Registration/check_email.html.twig', [
             'user' => $user,
-        ));
+        ]);
     }
 
     /**
      * Receive the confirmation token from user email provider, login the user.
      *
-     * @param Request $request
-     * @param string  $token
+     * @param string $token
      *
      * @return Response
      */
@@ -169,10 +166,10 @@ class RegistrationController extends Controller
             throw new AccessDeniedException('This user does not have access to this section.');
         }
 
-        return $this->render('@FOSUser/Registration/confirmed.html.twig', array(
+        return $this->render('@FOSUser/Registration/confirmed.html.twig', [
             'user' => $user,
             'targetUrl' => $this->getTargetUrlFromSession($request->getSession()),
-        ));
+        ]);
     }
 
     /**