+++ Form/Factory/FormFactory.php
@@ -38,12 +38,9 @@ class FormFactory implements FactoryInterface
     /**
      * FormFactory constructor.
      *
-     * @param FormFactoryInterface $formFactory
-     * @param string               $name
-     * @param string               $type
-     * @param array                $validationGroups
+     * @param array $validationGroups
      */
-    public function __construct(FormFactoryInterface $formFactory, $name, $type, array $validationGroups = null)
+    public function __construct(FormFactoryInterface $formFactory, string $name, string $type, array $validationGroups = null)
     {
         $this->formFactory = $formFactory;
         $this->name = $name;
@@ -54,9 +51,9 @@ class FormFactory implements FactoryInterface
     /**
      * {@inheritdoc}
      */
-    public function createForm(array $options = array())
+    public function createForm(array $options = [])
     {
-        $options = array_merge(array('validation_groups' => $this->validationGroups), $options);
+        $options = array_merge(['validation_groups' => $this->validationGroups], $options);
 
         return $this->formFactory->createNamed($this->name, $this->type, null, $options);
     }