+++ Model/UserInterface.php
@@ -14,10 +14,7 @@ namespace FOS\UserBundle\Model;
 use Symfony\Component\Security\Core\User\EquatableInterface;
 use Symfony\Component\Security\Core\User\UserInterface as BaseUserInterface;
 
-/**
- * @internal Only for back compatibility. Remove / merge when dropping support for Symfony 4
- */
-interface FosUserInterface extends \Serializable
+interface UserInterface extends BaseUserInterface, EquatableInterface, \Serializable
 {
     const ROLE_DEFAULT = 'ROLE_USER';
 
@@ -161,8 +158,6 @@ interface FosUserInterface extends \Serializable
     /**
      * Sets the timestamp that the user requested a password reset.
      *
-     * @param \DateTime|null $date
-     *
      * @return static
      */
     public function setPasswordRequestedAt(\DateTime $date = null);
@@ -178,8 +173,6 @@ interface FosUserInterface extends \Serializable
 
     /**
      * Sets the last login time.
-     *
-     * @param \DateTime|null $time
      *
      * @return static
      */
@@ -203,8 +196,6 @@ interface FosUserInterface extends \Serializable
      * Sets the roles of the user.
      *
      * This overwrites any previous roles.
-     *
-     * @param array $roles
      *
      * @return static
      */
@@ -275,26 +266,4 @@ interface FosUserInterface extends \Serializable
      * @see DisabledException
      */
     public function isEnabled();
-}
-
-// This is required to support apps that explicitly check if a user is an instance of AdvancedUserInterface
-if (interface_exists('\Symfony\Component\Security\Core\User\AdvancedUserInterface')) {
-    /**
-     * @author Thibault Duplessis <thibault.duplessis@gmail.com>
-     * @author Johannes M. Schmitt <schmittjoh@gmail.com>
-     *
-     * @deprecated since Symfony 4.1. Remove in Nov 2023 (End of support for security fixes SF 4.4)
-     */
-    interface UserInterface extends FosUserInterface, \Symfony\Component\Security\Core\User\AdvancedUserInterface
-    {
-    }
-} else {
-    /**
-     * @author Thibault Duplessis <thibault.duplessis@gmail.com>
-     * @author Johannes M. Schmitt <schmittjoh@gmail.com>
-     * @author Julian Finkler <julian@developer-heaven.de>
-     */
-    interface UserInterface extends FosUserInterface, BaseUserInterface, EquatableInterface
-    {
-    }
 }