+++ Changelog.md
@@ -1,9 +1,18 @@
 Changelog
 =========
 
-### 2.1.3 (2018-xx-xx)
+### 3.0.0 (xxxx-xx-xx)
 
-* Added Mongolian translation.
+* Dropped PHP 5 and PHP 7.1 support,
+* Switched to short array syntax,
+* Dropped Symfony 2.8 support,
+* Bumped minimum Symfony 3 requirement to 3.4,
+* Bumped minimum Symfony 4 requirement to 4.3,
+* All controllers have been marked final,
+* All controllers extend `AbstractController` instead of `Controller`,
+* Replaced EngineInterface type hints by Twig\Environment,
+* UserInterface no longer implements AdvancedUserInterface,
+* Added Mongolian translation,
 
 ### 2.1.2 (2018-03-08)